Add Squidle client facade and transport.
